refactor(configs): move DischargeLevelsTabs class to discharge_levels.py for better organization
